fix: import sibling packages without building
