+++ packages/core/index.ts
@@ -1,0 +1,3 @@
+export { GardenJS } from './src/garden';
+export { SwapperActions, SwapperRole } from './src/swapper';
+export { catalogWalletActions } from './src/catalogActions';
