--- packages/core/src/index.ts
@@ -1,3 +1,0 @@
-export { GardenJS } from "./lib/garden";
-export { SwapperActions, SwapperRole } from "./lib/swapper";
-export { catalogWalletActions } from "./lib/catalogActions";
