+++ packages/core/tsconfig.spec.json
@@ -6,6 +6,8 @@
     "types": ["node", "chrome"]
   },
   "include": [
+    "src/**/*.ts",
+    "index.ts",
     "jest.config.ts",
     "src/**/*.test.ts",
     "src/**/*.spec.ts",
