+++ tsconfig.json
@@ -1,16 +1,17 @@
 {
-    "compilerOptions": {
-        "composite": true,
-        "esModuleInterop": true,
-        "forceConsistentCasingInFileNames": true,
-        "lib": ["DOM", "ES2020"],
-        "module": "CommonJS",
-        "paths": {
-            "gardenfi/*": ["./packages/*/src"]
-        },
-        "moduleResolution": "node",
-        "resolveJsonModule": true,
-        "strict": true,
-        "target": "ES2020"
-    }
+  "compilerOptions": {
+    "composite": true,
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "lib": ["DOM", "ES2020"],
+    "module": "CommonJS",
+    "baseUrl": ".",
+    "paths": {
+      "@gardenfi/*": ["./packages/*"]
+    },
+    "moduleResolution": "node",
+    "resolveJsonModule": true,
+    "strict": true,
+    "target": "ES2020"
+  }
 }
